app/requirements.txt
flask==3.0.0
phe==1.5.0
pyzmq==22.3.0
